+++ lib/circleci/coverage_reporter/github_client.rb
@@ -1,4 +1,5 @@
 require_relative 'abstract_vcs_client'
+require_relative 'errors'
 
 module CircleCI
   module CoverageReporter
@@ -6,7 +7,18 @@ module CircleCI
       # @note Implement {AbstractVCSClient#create_comment}
       # @param reports [Array<Report>]
       # @return [void]
+      # @raise [RequestError]
       def create_comment(reports)
+        resp = request(reports)
+        body = JSON.parse(resp.body)
+        raise RequestError.new(body['message'], resp) unless resp.success?
+      end
+
+      private
+
+      # @param reports [Array<Report>]
+      # @return [Faraday::Response]
+      def request(reports)
         Faraday.new(url: 'https://api.github.com').post do |req|
           req.url ['/repos', configuration.project, 'commits', configuration.current_revision, 'comments'].join('/')
           req.headers['Authorization'] = "token #{token}"
@@ -14,8 +26,6 @@ module CircleCI
           req.body = JSON.generate(body: reports.join("\n"))
         end
       end
-
-      private
 
       # @return [Client]
       def configuration