+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Fixed
+- Raise `RequestError` if some API reqeust fails
+
 ## [0.1.2] - 2017-03-12
 ### Fixed
 
@@ -19,5 +23,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Initial release
 
+[Unreleased]: https://github.com/increments/circleci-coverage_reporter/compare/v0.1.2...HEAD
 [0.1.2]: https://github.com/increments/circleci-coverage_reporter/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/increments/circleci-coverage_reporter/compare/v0.1.0...v0.1.1